Module pill_engine::game

Macros

define_new_pill_slotmap_key

Structs

AudioListenerComponent
The AudioListenerComponent type, a component responsible for enabling certaing listener to fetch position from

AudioManagerComponent
The AudioManagerComponent type, global component responsible for

AudioSourceComponent

CameraComponent

ComponentStorage

Engine
Heart of Pill Engine

EntityHandle

Error
The Error type, a wrapper around a dynamic error type.

GlobalComponentStorage

InputComponent

Material

MaterialHandle

Mesh

MeshHandle

MeshRenderingComponent

ResourceStorage

SceneHandle

Sound

Texture

TextureHandle

TimeComponent

TransformComponent

Enums

CameraAspectRatio

ResourceLoadType

SoundType

TextureType

Traits

Component

Context
Provides the context method for Result.

GlobalComponent

PillGame
Engine <-> Game interface

PillTypeMapKey
PillTypeMapKey is used to declare key types that are eligible for use with PillTypeMap.

Resource

Type Definitions

Color

KeyboardKey

MouseButton

Result
Result<T, Error>

Vector2f

Vector2i

Vector3f

Vector3i