Struct pill_engine::game::Engine

pub struct Engine {
    pub(crate) config: Config,
    pub(crate) game: Option<Box<dyn PillGame>>,
    pub(crate) renderer: Box<dyn PillRenderer>,
    pub(crate) scene_manager: SceneManager,
    pub(crate) system_manager: SystemManager,
    pub(crate) resource_manager: ResourceManager,
    pub(crate) global_components: PillTypeMap,
    pub(crate) input_queue: VecDeque<InputEvent>,
    pub(crate) render_queue: Vec<RenderQueueItem>,
    pub(crate) window_size: PhysicalSize<u32>,
    pub(crate) frame_delta_time: f32,
}

Heart of Pill Engine

Fields
config: Configgame: Option<Box<dyn PillGame>>renderer: Box<dyn PillRenderer>scene_manager: SceneManagersystem_manager: SystemManagerresource_manager: ResourceManagerglobal_components: PillTypeMapinput_queue: VecDeque<InputEvent>render_queue: Vec<RenderQueueItem>window_size: PhysicalSize<u32>frame_delta_time: f32
Implementations

impl Engine

Pill Engine internal functions

fn create_default_resources(&mut self) -> Result<()>

impl Engine

Pill Engine internal API

pub fn new(
game: Box<dyn PillGame>,
renderer: Box<dyn PillRenderer>,
config: Config
) -> Self

pub fn initialize(&mut self, window_size: PhysicalSize<u32>) -> Result<()>

Initializes Pill Engine

Creates default global components, adds default systems, creates default resources, initializes game

pub fn update(&mut self, delta_time: Duration)

Main engine update function

Runs all systems in order: PreGame -> Game -> PostGame

pub fn shutdown(&mut self)

pub fn resize(&mut self, new_window_size: PhysicalSize<u32>)

pub fn pass_keyboard_key_input(&mut self, keyboard_input: &KeyboardInput)

pub fn pass_mouse_key_input(&mut self, key: &MouseButton, state: &ElementState)

pub fn pass_mouse_wheel_input(&mut self, delta: &MouseScrollDelta)

pub fn pass_mouse_delta_input(&mut self, delta: &(f64, f64))

pub fn pass_mouse_position_input(&mut self, position: &PhysicalPosition<f64>)

pub fn get_input_queue(&self) -> &VecDeque<InputEvent>

impl Engine

Pill Engine game API

pub fn add_system(
&mut self,
name: &str,
system_function: fn(engine: &mut Engine) -> Result<()>
) -> Result<()>

Adds game-defined system to the game update phase

pub fn remove_system(&mut self, name: &str) -> Result<()>

Removes game-defined system

pub fn toggle_system(&mut self, name: &str, enabled: bool) -> Result<()>

Toggles game-defined system

pub fn build_entity(&mut self, scene_handle: SceneHandle) -> EntityBuilder<'_>

Returns EntityBuilder, allowing for handy entity creation

pub fn create_entity(
&mut self,
scene_handle: SceneHandle
) -> Result<EntityHandle>

pub fn remove_entity(
&mut self,
entity_handle: EntityHandle,
scene_handle: SceneHandle
) -> Result<()>

pub fn register_component<T>(&mut self, scene_handle: SceneHandle) -> Result<()> where
T: Component<Storage = ComponentStorage<T>>,

Registers new component type in scene specified with scene handle

pub fn add_component_to_entity<T>(
&mut self,
scene_handle: SceneHandle,
entity_handle: EntityHandle,
component: T
) -> Result<()> where
T: Component<Storage = ComponentStorage<T>>,

Adds new component to the entity specified with scene and entity handle

pub fn remove_component_from_entity<T>(
&mut self,
scene_handle: SceneHandle,
entity_handle: EntityHandle
) -> Result<()> where
T: Component<Storage = ComponentStorage<T>>,

Removes component from the entity specified with scene and entity handle

pub fn get_entity_component<T>(
&self,
entity_handle: EntityHandle,
scene_handle: SceneHandle
) -> Result<&RefCell<Option<T>>> where
T: Component<Storage = ComponentStorage<T>>,

pub fn add_global_component<T>(&mut self, component: T) -> Result<()> where
T: GlobalComponent<Storage = GlobalComponentStorage<T>>,

Adds global component to engine

pub fn get_global_component<T>(&self) -> Result<&T> where
T: GlobalComponent<Storage = GlobalComponentStorage<T>>,

Returns global component

pub fn get_global_component_mut<T>(&mut self) -> Result<&mut T> where
T: GlobalComponent<Storage = GlobalComponentStorage<T>>,

Returns global mutable component

pub fn remove_global_component<T>(&mut self) -> Result<()> where
T: GlobalComponent<Storage = GlobalComponentStorage<T>>,

Removes global component from the engine

pub fn iterate_one_component<A>(
&self
) -> Result<impl Iterator<Item = &RefCell<Option<A>>>> where
A: Component<Storage = ComponentStorage<A>>,

Returns iterator for specified component

pub fn iterate_one_component_with_entities<A>(
&self
) -> Result<impl Iterator<Item = (EntityHandle, &RefCell<Option<A>>)>> where
A: Component<Storage = ComponentStorage<A>>,

Returns iterator for specified component

Additionally returns entity handle to matching entities

pub fn iterate_two_components<A, B>(
&self
) -> Result<impl Iterator<Item = (&RefCell<Option<A>>, &RefCell<Option<B>>)>> where
A: Component<Storage = ComponentStorage<A>>,
B: Component<Storage = ComponentStorage<B>>,

Returns iterator for specified component pair

Iterator fetches specified components only for those entities which have them all

pub fn iterate_two_components_with_entities<A, B>(
&self
) -> Result<impl Iterator<Item = (EntityHandle, &RefCell<Option<A>>, &RefCell<Option<B>>)>> where
A: Component<Storage = ComponentStorage<A>>,
B: Component<Storage = ComponentStorage<B>>,

Returns iterator for specified component pair

Iterator fetches specified components only for those entities which have them all Additionally returns entity handle to matching entities

pub fn iterate_three_components<A, B, C>(
&self
) -> Result<impl Iterator<Item = (&RefCell<Option<A>>, &RefCell<Option<B>>, &RefCell<Option<C>>)>> where
A: Component<Storage = ComponentStorage<A>>,
B: Component<Storage = ComponentStorage<B>>,
C: Component<Storage = ComponentStorage<C>>,

Returns iterator for specified component triple

Iterator fetches specified components only for those entities which have them all

pub fn iterate_three_components_with_entities<A, B, C>(
&self
) -> Result<impl Iterator<Item = (EntityHandle, &RefCell<Option<A>>, &RefCell<Option<B>>, &RefCell<Option<C>>)>> where
A: Component<Storage = ComponentStorage<A>>,
B: Component<Storage = ComponentStorage<B>>,
C: Component<Storage = ComponentStorage<C>>,

Returns iterator for specified component triple

Iterator fetches specified components only for those entities which have them all Additionally returns entity handle to matching entities

pub fn iterate_four_components<A, B, C, D>(
&self
) -> Result<impl Iterator<Item = (&RefCell<Option<A>>, &RefCell<Option<B>>, &RefCell<Option<C>>, &RefCell<Option<D>>)>> where
A: Component<Storage = ComponentStorage<A>>,
B: Component<Storage = ComponentStorage<B>>,
C: Component<Storage = ComponentStorage<C>>,
D: Component<Storage = ComponentStorage<D>>,

Returns iterator for specified component quadruple

Iterator fetches specified components only for those entities which have them all

pub fn iterate_four_components_with_entities<A, B, C, D>(
&self
) -> Result<impl Iterator<Item = (EntityHandle, &RefCell<Option<A>>, &RefCell<Option<B>>, &RefCell<Option<C>>, &RefCell<Option<D>>)>> where
A: Component<Storage = ComponentStorage<A>>,
B: Component<Storage = ComponentStorage<B>>,
C: Component<Storage = ComponentStorage<C>>,
D: Component<Storage = ComponentStorage<D>>,

Returns iterator for specified component quadruple

Iterator fetches specified components only for those entities which have them all Additionally returns entity handle to matching entities

pub fn create_scene(&mut self, name: &str) -> Result<SceneHandle>

pub fn get_scene_handle(&self, name: &str) -> Result<SceneHandle>

Returns handle to the scene specified by its name

pub fn set_active_scene(&mut self, scene_handle: SceneHandle) -> Result<()>

pub fn get_active_scene_handle(&self) -> Result<SceneHandle>

Returns handle to the active scene

pub fn remove_scene(&mut self, scene_handle: SceneHandle) -> Result<()>

pub fn register_resource_type<T>(
&mut self,
max_resource_count: usize
) -> Result<()> where
T: Resource<Storage = ResourceStorage<T>>,

pub fn add_resource<T>(&mut self, resource: T) -> Result<T::Handle> where
T: Resource<Storage = ResourceStorage<T>>,

pub fn get_resource<'a, T>(
&'a self,
resource_handle: &'a T::Handle
) -> Result<&'a T> where
T: Resource<Storage = ResourceStorage<T>>,

pub fn get_resource_by_name<T>(&self, name: &str) -> Result<&T> where
T: Resource<Storage = ResourceStorage<T>>,

Returns resource specified by its name

pub fn get_resource_handle<T>(&self, name: &str) -> Result<T::Handle> where
T: Resource<Storage = ResourceStorage<T>>,

Returns handle to resource specified by the name of this resource

pub fn get_resource_mut<'a, T>(
&'a mut self,
resource_handle: &'a T::Handle
) -> Result<&'a mut T> where
T: Resource<Storage = ResourceStorage<T>>,

pub fn get_resource_by_name_mut<T>(&mut self, name: &str) -> Result<&mut T> where
T: Resource<Storage = ResourceStorage<T>>,

Returns mutable resource specified by its name

pub fn remove_resource<T>(&mut self, resource_handle: &T::Handle) -> Result<()> where
T: Resource<Storage = ResourceStorage<T>>,

pub fn remove_resource_by_name<T>(&mut self, name: &str) -> Result<()> where
T: Resource<Storage = ResourceStorage<T>>,