Trait pill_engine::game::PillTypeMapKey

pub trait PillTypeMapKey: Any {
    type Storage: Send;
}

PillTypeMapKey is used to declare key types that are eligible for use with PillTypeMap.

Associated Types

type Storage: Send

Defines the value type that corresponds to this PillTypeMapKey.

Loading content...